Rename publishers/*/data to publishers/*/repositories.
