+++ src/app/pages/home/home.page.ts
@@ -5,8 +5,8 @@ import { combineLatest, of, zip } from 'rxjs';
 import { concatMap, map, tap } from 'rxjs/operators';
 import { CameraService } from 'src/app/services/camera/camera.service';
 import { CollectorService } from 'src/app/services/collector/collector.service';
-import { AssetRepository } from 'src/app/services/publisher/numbers-storage/data/asset/asset-repository.service';
 import { NumbersStorageApi } from 'src/app/services/publisher/numbers-storage/numbers-storage-api.service';
+import { AssetRepository } from 'src/app/services/publisher/numbers-storage/repositories/asset/asset-repository.service';
 import { PublishersAlert } from 'src/app/services/publisher/publishers-alert/publishers-alert.service';
 import { getOldProof } from 'src/app/services/repositories/proof/old-proof-adapter';
 import { ProofRepository } from 'src/app/services/repositories/proof/proof-repository.service';
