+++ src/app/services/publisher/numbers-storage/repositories/asset/asset.ts
@@ -6,7 +6,6 @@ export interface Asset extends Tuple {
   readonly proof_hash: string;
   readonly owner: string;
   readonly asset_file: string;
-  readonly asset_file_thumbnail: string;
   readonly information: SortedProofInformation;
   readonly signature: OldSignature[];
   readonly caption: string;
