+++ src/app/shared/post-capture-card/post-capture-card.component.spec.ts
@@ -6,8 +6,8 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatIconTestingModule } from '@angular/material/icon/testing';
 import { MatListModule } from '@angular/material/list';
 import { IonicModule } from '@ionic/angular';
-import { Asset } from 'src/app/services/publisher/numbers-storage/data/asset/asset';
 import { Transaction } from 'src/app/services/publisher/numbers-storage/numbers-storage-api.service';
+import { Asset } from 'src/app/services/publisher/numbers-storage/repositories/asset/asset';
 import { getTranslocoModule } from 'src/app/transloco/transloco-root.module.spec';
 import { PostCaptureCardComponent } from './post-capture-card.component';
 
@@ -19,7 +19,6 @@ describe('PostCaptureCardComponent', () => {
     proof_hash: 'abcdef1234567890',
     owner: 'me',
     asset_file: 'https://picsum.photos/200/300',
-    asset_file_thumbnail: 'https://picsum.photos/200/300',
     information: { proof: { hash: '', timestamp: 0, mimeType: 'image/jpeg' }, information: [] },
     signature: [],
     caption: '',
@@ -29,7 +28,11 @@ describe('PostCaptureCardComponent', () => {
   const expectedTranasction: Transaction = {
     id: '',
     sender: '',
-    asset: expectedAsset,
+    asset: {
+      asset_file_thumbnail: 'https://picsum.photos/200/300',
+      caption: '',
+      id: 'abcd-efgh-ijkl'
+    },
     created_at: '',
     expired: false,
     fulfilled_at: ''
